+++ src/components/AppContainer.tsx
@@ -18,7 +18,7 @@ const appBrandVariants: BrandVariants = {
   30: "#e1d7ff",
   40: "#cec1fa",
   50: "#b9aaf5",
-  60: "#8a79e3",
+  60: "#9e90e8",
   70: "#8775e3",
   80: "#6f5ed9",
   90: "#5847cf",
@@ -80,6 +80,7 @@ const customLightTheme: Theme = {
   colorNeutralStroke2: "#e7ddff", // Your lavender for subtle borders
   colorNeutralStroke3: "#ede6ff",
   colorNeutralStrokeAccessible: "#6943d1", // Accessible border color
+  colorNeutralStrokeAccessibleHover: "#8361e0",
   colorNeutralStrokeDisabled: "#e7ddff",
   colorNeutralStrokeOnBrand: "#ffffff",
