+++ src/components/RecipePage/RecipeContent/RecipeContent.styles.ts
@@ -1,0 +1,99 @@
+import { tokens, makeStyles, shorthands } from "@fluentui/react-components";
+
+export const useStyles = makeStyles({
+  recipeContent: {
+    lineHeight: "1.8",
+    fontSize: "17px",
+    fontFamily: "'Georgia', serif",
+    color: tokens.colorNeutralForeground1,
+    "& h2": {
+      marginTop: "32px",
+      marginBottom: "16px",
+      fontWeight: "600",
+      borderBottom: `1px solid ${tokens.colorNeutralStroke2}`,
+      paddingBottom: "10px",
+      fontSize: "1.5rem",
+    },
+    "& ul, & ol": {
+      paddingLeft: "25px",
+      margin: "18px 0",
+    },
+    "& li": {
+      marginBottom: "12px",
+      position: "relative",
+    },
+    "& p": {
+      margin: "18px 0",
+    },
+    "& a": {
+      color: tokens.colorBrandForegroundLink,
+      textDecoration: "none",
+      borderBottom: `1px solid ${tokens.colorBrandStroke2Hover}`,
+      transition: "border-bottom 0.2s ease",
+      "&:hover": {
+        borderBottomWidth: "2px",
+      },
+    },
+    "& img": {
+      maxWidth: "100%",
+      ...shorthands.borderRadius("8px"),
+      marginTop: "16px",
+      marginBottom: "16px",
+    },
+    "& blockquote": {
+      borderLeft: `4px solid ${tokens.colorBrandStroke1}`,
+      paddingLeft: "16px",
+      margin: "20px 0",
+      color: tokens.colorNeutralForeground2,
+      fontStyle: "italic",
+    },
+  },
+  contentTextarea: {
+    width: "100%",
+    minHeight: "350px",
+    fontFamily: "'Georgia', serif",
+    fontSize: "17px",
+    lineHeight: "1.8",
+    ...shorthands.padding("20px"),
+    boxSizing: "border-box",
+    ...shorthands.borderRadius("8px"),
+    border: `1px solid ${tokens.colorNeutralStroke1}`,
+    backgroundColor: tokens.colorNeutralBackground1,
+    color: tokens.colorNeutralForeground1,
+    resize: "vertical",
+    transition: "all 0.2s ease",
+    marginBottom: "24px",
+    boxShadow: "inset 0 2px 4px rgba(0,0,0,0.02)",
+    "&:focus": {
+      ...shorthands.borderColor(tokens.colorBrandStroke1),
+      outline: "none",
+      boxShadow: `0 0 0 2px ${tokens.colorBrandStroke2Hover}20`,
+    },
+    "&::placeholder": {
+      color: tokens.colorNeutralForeground3,
+      fontStyle: "italic",
+    },
+  },
+  buttonRow: {
+    display: "flex",
+    justifyContent: "flex-end",
+    ...shorthands.gap("12px"),
+    marginTop: "24px",
+  },
+  editModeIndicator: {
+    backgroundColor: tokens.colorBrandBackground,
+    color: tokens.colorNeutralForegroundInverted,
+    ...shorthands.padding("6px", "12px"),
+    ...shorthands.borderRadius("4px"),
+    fontSize: "0.875rem",
+    fontWeight: "500",
+    position: "absolute",
+    top: "20px",
+    right: "20px",
+    zIndex: 10,
+    display: "flex",
+    alignItems: "center",
+    gap: "6px",
+    boxShadow: "rgba(0, 0, 0, 0.1) 0px 2px 8px",
+  },
+});
