+++ src/components/RecipePage/RecipePage.styles.ts
@@ -16,9 +16,9 @@ export const useStyles = makeStyles({
     ...shorthands.padding("0", "0", "30px"),
     display: "flex",
     flexDirection: "column",
-    ...shorthands.gap("30px"),
+    gap: "30px",
     backgroundColor: "white",
-    ...shorthands.borderRadius("16px"),
+    borderRadius: "16px",
     boxShadow: "rgba(0, 0, 0, 0.08) 0px 8px 24px",
     overflow: "hidden",
   },
@@ -31,54 +31,10 @@ export const useStyles = makeStyles({
       flexDirection: "column",
     },
   },
-  imageContainer: {
-    width: "45%",
-    height: "380px",
-    position: "relative",
-    "@media (max-width: 768px)": {
-      width: "100%",
-      height: "320px",
-      marginBottom: "20px",
-    },
-  },
-  imageUploadContainer: {
-    width: "45%",
-    height: "380px",
-    position: "relative",
-    display: "flex",
-    flexDirection: "column",
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: tokens.colorNeutralBackground1,
-    border: `2px dashed ${tokens.colorNeutralStrokeAccessible}`,
-    ...shorthands.borderRadius("12px"),
-    cursor: "pointer",
-    ...shorthands.gap("16px"),
-    ...shorthands.margin("16px", "0", "16px", "16px"),
-    transition: "all 0.2s ease-in-out",
-    "&:hover": {
-      backgroundColor: tokens.colorNeutralBackground2,
-      ...shorthands.borderColor(tokens.colorBrandBackground),
-    },
-    "@media (max-width: 768px)": {
-      width: "100%",
-      height: "320px",
-      marginBottom: "20px",
-      margin: "16px",
-    },
-  },
-  uploadIcon: {
-    color: tokens.colorBrandForeground1,
-    fontSize: "48px",
-  },
-  uploadText: {
-    color: tokens.colorNeutralForeground2,
-    fontWeight: "500",
-  },
-  recipeImage: {
-    objectFit: "cover",
-    borderRadius: "12px",
-    boxShadow: "rgba(0, 0, 0, 0.06) 0px 2px 8px",
+  divider: {
+    padding: "0px 36px",
+    margin: "10px auto",
+    opacity: "0.6",
   },
   contentContainer: {
     display: "flex",
@@ -89,229 +45,5 @@ export const useStyles = makeStyles({
     "@media (max-width: 768px)": {
       padding: "0 24px",
     },
-  },
-  headerSection: {
-    display: "flex",
-    flexDirection: "column",
-    gap: "20px",
-  },
-  titleRow: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "flex-start",
-    gap: "16px",
-  },
-  title: {
-    margin: 0,
-    fontWeight: "600",
-    fontFamily: "'Georgia', serif",
-    fontSize: "2.2rem",
-    lineHeight: "1.2",
-  },
-  titleInput: {
-    fontSize: "2rem",
-    fontWeight: "600",
-    fontFamily: "'Georgia', serif",
-    width: "100%",
-    border: "none",
-    borderBottom: `2px solid ${tokens.colorNeutralStroke1}`,
-    ...shorthands.borderRadius("0"),
-    paddingBottom: "8px",
-    "&:focus": {
-      ...shorthands.borderColor(tokens.colorBrandStroke1),
-      outline: "none",
-    },
-  },
-  actionButtons: {
-    display: "flex",
-    ...shorthands.gap("12px"),
-    alignItems: "center",
-  },
-  actionButton: {
-    minWidth: "36px",
-    height: "36px",
-    ...shorthands.padding("0"),
-    ...shorthands.borderRadius("18px"),
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    transition: "transform 0.2s ease",
-    "&:hover": {
-      transform: "scale(1.05)",
-    },
-  },
-  date: {
-    color: tokens.colorNeutralForeground3,
-    marginTop: "4px",
-    fontStyle: "italic",
-  },
-  tagsContainer: {
-    display: "flex",
-    flexWrap: "wrap",
-    ...shorthands.gap("10px"),
-    marginTop: "12px",
-    alignItems: "center",
-  },
-  tag: {
-    backgroundColor: tokens.colorNeutralBackground3,
-    ...shorthands.borderRadius("20px"),
-    ...shorthands.padding("8px", "16px"),
-    fontSize: "0.875rem",
-    color: tokens.colorNeutralForeground2,
-    display: "inline-flex",
-    alignItems: "center",
-    gap: "6px",
-    fontWeight: "500",
-    transition: "all 0.2s ease",
-  },
-  editableTag: {
-    backgroundColor: tokens.colorBrandBackground2,
-    ...shorthands.borderRadius("20px"),
-    ...shorthands.padding("8px", "16px"),
-    fontSize: "0.875rem",
-    color: tokens.colorBrandForeground2,
-    display: "inline-flex",
-    alignItems: "center",
-    gap: "6px",
-    fontWeight: "500",
-    transition: "all 0.2s ease",
-    "&:hover": {
-      backgroundColor: tokens.colorBrandBackground,
-    },
-  },
-  removeTagIcon: {
-    cursor: "pointer",
-    fontSize: "12px",
-    opacity: "0.7",
-    "&:hover": {
-      opacity: "1",
-    },
-  },
-  tagInput: {
-    backgroundColor: tokens.colorNeutralBackground1,
-    ...shorthands.borderRadius("20px"),
-    ...shorthands.padding("8px", "16px"),
-    fontSize: "0.875rem",
-    width: "150px",
-    border: `1px solid ${tokens.colorNeutralStroke1}`,
-    "&:focus": {
-      ...shorthands.borderColor(tokens.colorBrandStroke1),
-      outline: "none",
-    },
-  },
-  addTagButton: {
-    ...shorthands.borderRadius("20px"),
-    width: "32px",
-    height: "32px",
-    display: "inline-flex",
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: tokens.colorBrandBackground,
-    color: tokens.colorNeutralForegroundInverted,
-    cursor: "pointer",
-    transition: "transform 0.2s ease, background-color 0.2s ease",
-    "&:hover": {
-      backgroundColor: tokens.colorBrandBackgroundHover,
-      transform: "scale(1.05)",
-    },
-  },
-  recipeContent: {
-    lineHeight: "1.8",
-    fontSize: "17px",
-    fontFamily: "'Georgia', serif",
-    color: tokens.colorNeutralForeground1,
-    "& h2": {
-      marginTop: "32px",
-      marginBottom: "16px",
-      fontWeight: "600",
-      borderBottom: `1px solid ${tokens.colorNeutralStroke2}`,
-      paddingBottom: "10px",
-      fontSize: "1.5rem",
-    },
-    "& ul, & ol": {
-      paddingLeft: "25px",
-      margin: "18px 0",
-    },
-    "& li": {
-      marginBottom: "12px",
-      position: "relative",
-    },
-    "& p": {
-      margin: "18px 0",
-    },
-    "& a": {
-      color: tokens.colorBrandForegroundLink,
-      textDecoration: "none",
-      borderBottom: `1px solid ${tokens.colorBrandStroke2Hover}`,
-      transition: "border-bottom 0.2s ease",
-      "&:hover": {
-        borderBottomWidth: "2px",
-      },
-    },
-    "& img": {
-      maxWidth: "100%",
-      ...shorthands.borderRadius("8px"),
-      marginTop: "16px",
-      marginBottom: "16px",
-    },
-    "& blockquote": {
-      borderLeft: `4px solid ${tokens.colorBrandStroke1}`,
-      paddingLeft: "16px",
-      margin: "20px 0",
-      color: tokens.colorNeutralForeground2,
-      fontStyle: "italic",
-    },
-  },
-  contentTextarea: {
-    width: "100%",
-    minHeight: "350px", // Reduced from 450px for better proportion
-    fontFamily: "'Georgia', serif",
-    fontSize: "17px",
-    lineHeight: "1.8",
-    ...shorthands.padding("20px"),
-    boxSizing: "border-box", // Ensures padding is included in width calculation
-    ...shorthands.borderRadius("8px"),
-    border: `1px solid ${tokens.colorNeutralStroke1}`,
-    backgroundColor: tokens.colorNeutralBackground1,
-    color: tokens.colorNeutralForeground1,
-    resize: "vertical",
-    transition: "all 0.2s ease",
-    marginBottom: "24px", // Add spacing at bottom
-    boxShadow: "inset 0 2px 4px rgba(0,0,0,0.02)",
-    "&:focus": {
-      ...shorthands.borderColor(tokens.colorBrandStroke1),
-      outline: "none",
-      boxShadow: `0 0 0 2px ${tokens.colorBrandStroke2Hover}20`, // Subtle focus ring
-    },
-    "&::placeholder": {
-      color: tokens.colorNeutralForeground3,
-      fontStyle: "italic",
-    },
-  },
-  divider: {
-    margin: "10px 36px",
-    opacity: "0.6",
-  },
-  buttonRow: {
-    display: "flex",
-    justifyContent: "flex-end",
-    ...shorthands.gap("12px"),
-    marginTop: "24px",
-  },
-  editModeIndicator: {
-    backgroundColor: tokens.colorBrandBackground,
-    color: tokens.colorNeutralForegroundInverted,
-    ...shorthands.padding("6px", "12px"),
-    ...shorthands.borderRadius("4px"),
-    fontSize: "0.875rem",
-    fontWeight: "500",
-    position: "absolute",
-    top: "20px",
-    right: "20px",
-    zIndex: 10,
-    display: "flex",
-    alignItems: "center",
-    gap: "6px",
-    boxShadow: "rgba(0, 0, 0, 0.1) 0px 2px 8px",
   },
 });
