+++ src/context/SearchBoxProvider/SearchBoxProvider.tsx
@@ -9,3 +9,4 @@ export const SearchBoxContext = React.createContext<SearchBoxContextValue>({
 export const SearchBoxProvider = SearchBoxContext.Provider;
 
 export const useSearchBox = () => React.useContext(SearchBoxContext);
+1;
