+++ src/context/index.ts
@@ -1,1 +1,2 @@
 export * from "./SearchBoxProvider";
+export * from "./RecipeProvider";
